% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-analyses.R
\name{all_analyses}
\alias{all_analyses}
\title{Outputs the following:
\enumerate{
\item a three tab (.xlsx) spreadsheet with length-adjusted averages
\item a six tab (.xlsx) spreadsheet with summary tables
\item bar plots in (.jpeg) image files for mercury results in sport fish and prey fish
\item scatterplots in (.jpeg) image files for every sport fish species present in the dataset with mercury results
}}
\usage{
all_analyses(filename)
}
\arguments{
\item{filename}{Filename (or filepath, if necessary) for CEDEN AQT download data (.tsv) file.}
}
\description{
Outputs the following:
\enumerate{
\item a three tab (.xlsx) spreadsheet with length-adjusted averages
\item a six tab (.xlsx) spreadsheet with summary tables
\item bar plots in (.jpeg) image files for mercury results in sport fish and prey fish
\item scatterplots in (.jpeg) image files for every sport fish species present in the dataset with mercury results
}
}
