% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length-adjusted-averages.R
\name{length_adjusted_averages}
\alias{length_adjusted_averages}
\title{Outputs a three tab (.xlsx) spreadsheet with length-adjusted averages, length-adjusted averages with additional summary calculations, and intermediate length-adjusted result calculations}
\usage{
length_adjusted_averages(filename)
}
\arguments{
\item{filename}{Filename (or filepath, if necessary) for CEDEN AQT download data (.tsv) file.}
}
\description{
Outputs a three tab (.xlsx) spreadsheet with length-adjusted averages, length-adjusted averages with additional summary calculations, and intermediate length-adjusted result calculations
}
