% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury-by-species-barplot.R
\name{mercury_by_species_barplot}
\alias{mercury_by_species_barplot}
\title{Outputs a (.jpeg file) bar plot for mercury concentration by species, for a given source file and type of fish (sport or prey)}
\usage{
mercury_by_species_barplot(filename, fish_class)
}
\arguments{
\item{filename}{Filename (or filepath, if necessary) for CEDEN AQT download data (.tsv) file.}

\item{fish_class}{Type of fish (either "sport" or "prey")}
}
\description{
Outputs a (.jpeg file) bar plot for mercury concentration by species, for a given source file and type of fish (sport or prey)
}
