% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mercury-vs-length-scatterplot.R
\name{mercury_vs_length_scatterplot}
\alias{mercury_vs_length_scatterplot}
\title{Outputs a (.jpeg file) scatterplot of fish mercury concentration vs fish length in (mm) for a given sourcefile and species}
\usage{
mercury_vs_length_scatterplot(filename, species_common_name)
}
\arguments{
\item{filename}{Filename and realtive path for CEDEN AQT download data (.tsv) file.}

\item{species_common_name}{common name of fish species to be plotted (CommonName field in the CEDEN AQT download)}
}
\description{
Outputs a (.jpeg file) scatterplot of fish mercury concentration vs fish length in (mm) for a given sourcefile and species
}
