% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-tables.R
\name{summary_tables}
\alias{summary_tables}
\title{Outputs a six tab (.xlsx) spreadsheet with summary tables for your dataset}
\usage{
summary_tables(filename)
}
\arguments{
\item{filename}{Filename (or filepath, if necessary) for CEDEN AQT download data (.tsv) file.}
}
\description{
Outputs a six tab (.xlsx) spreadsheet with summary tables for your dataset
}
